LayoutTests/platform/blackberry/editing/undo/undo-smart-delete-word-expected.txt

CONSOLE MESSAGE: line 28: [object Text]
CONSOLE MESSAGE: line 29: 3
CONSOLE MESSAGE: line 30: [object Text]
CONSOLE MESSAGE: line 31: 3
Tests: 
Select a word via double-click. Delete. Then undo the delete. The space that got smart deleted should now be selected.
foo bar baz
FAILED
